# perf(snowflake): improve audit log sql query