## Summary

This PR improves Snowflake ingestion performance by pushing down database pattern filtering to the audit log SQL query instead of filtering in Python after data retrieval.

### Key Changes

- Push down database allow/deny pattern to the audit log sql query
- Feature is configurable via `push_down_database_pattern_query_log` config flag (default: `true`)

### Testing

#### Functional Test
In a recipe with:

```yaml
database_pattern:
  allow:
    - '^SMOKE_TEST_DB$'
    - '^SMOKE_TEST_DB_2$'
```

The following filter was added to the SQL query:

```sql
AND (database_name RLIKE '^SMOKE_TEST_DB$' OR database_name RLIKE '^SMOKE_TEST_DB_2$')
```

Query results reduced from 135 rows to 71 rows, demonstrating effective filtering at the database level.

#### Performance Test
Significant improvements in query processing and overall ingestion performance:

**Before optimization:**
- Total records written: 1,713,778
- Duration: 3 hr 43 min
- Query log results: 2,237,425
- Queries extraction stage: 12,376.22s
- Query log fetch timer: 3,232.798s
- Audit log load timer: 6,782.747s

**After optimization:**
- Total records written: 111,807 (93% reduction)
- Duration: 33 min 41s (85% reduction)
- Query log results: 43,271 (98% reduction)
- Queries extraction stage: 652.06s (95% reduction)
- Query log fetch timer: 261.022s (92% reduction)
- Audit log load timer: 32.508s (99% reduction)